backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-celery-beat==2.5.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
Pillow==10.4.0
python-decouple==3.8
python-dotenv==1.0.0
cryptography==42.0.5
pyotp==2.9.0
qrcode==7.4.2
openpyxl==3.1.2
pandas==2.2.2
reportlab==4.0.9
sendgrid==6.11.0
twilio==8.10.3
python-dateutil==2.9.0
pytz==2024.1
django-extensions==3.2.3
django-debug-toolbar==4.3.0
django-filter==23.5
django-import-export==3.3.1
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-environ==0.11.2
gunicorn==21.2.0
whitenoise==6.6.0
drf-spectacular==0.27.0

# WebSocket Support
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Enterprise Security & Analytics
bcrypt==4.1.2
numpy==1.25.2
scikit-learn==1.3.2
joblib==1.3.2  # Secure model serialization (replaces pickle)
requests==2.32.3
djangorestframework-api-key==3.0.0
django-ratelimit==4.1.0
django-oauth-toolkit==1.7.1
sentry-sdk==1.38.0

# Monitoring & Health Checks
prometheus-client==0.19.0
psutil==5.9.6

# Distributed Tracing & APM
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-django==0.43b0
opentelemetry-instrumentation-requests==0.43b0
opentelemetry-instrumentation-redis==0.43b0
opentelemetry-instrumentation-psycopg2==0.43b0
opentelemetry-exporter-otlp==1.22.0
datadog==0.49.0
ddtrace==2.6.0

# Security & Compliance
django-axes==6.1.1
django-cryptography==1.1
python-gnupg==0.5.2
bandit==1.7.6
safety==3.0.1

# Advanced Features
python-magic==0.4.27
bleach==6.1.0
markdown==3.5.1

# Testing & Performance
coverage==7.3.2
locust==2.17.0
django-redis==5.4.0
faker==20.1.0  # For generating test data

# Logging
python-json-logger==2.0.7

# File validation
python-magic==0.4.27  # For secure file content validation

# AI & Machine Learning (New Features)
openai==1.3.0  # For AI content generation
textblob==0.17.1  # For sentiment analysis (optional)
